// src/interfaces/Autotest.ts
export interface Autotest {
    automatedTestClass: string
    automatedTestName: string
    comments: string
    setToIgnore: string
    testDuration: string
    testResult: string
    testPackage: string
    testLinkId: string
    priority: string
    generalWorkflow: string
    specificWorkflow: string
    partialCoverageOnly: string
}